Add Git Script to Count Commits from Branch
Add git script to count the number of new commits on a branch compared
to another branch.  The default branch for comparison is master.  The
goal of this is script is to make it easier to interactive rebase a
feature branch with master.  By knowing the number of commits the
feature branch deviates from master, one knows the number of commits to
rebase from HEAD.